Move common IDevID functionality out of EST views
