+++ trustpoint/est/tests/test_est_interface.py
@@ -4,11 +4,12 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 from cryptography import x509
+from devices.models import IssuedCredentialModel
 from django.test import RequestFactory
-from pki.util.x509 import CertificateGenerator
+from pki.util.idevid import IDevIDAuthenticator
+from pki.util.x509 import ClientCertificateAuthenticationError
 
 from est.views import (
-    ClientCertificateAuthenticationError,
     EstCACertsView,
     EstSimpleEnrollmentView,
     LoggedHttpResponse,
@@ -47,7 +48,7 @@ def get_mock_truststore(certificates: list[x509.Certificate]) -> MagicMock:
 
 @patch('pki.models.credential.CredentialModel.objects.filter')
 @patch('devices.models.IssuedCredentialModel.objects.get')
-def test_get_credential_for_certificate(mock_get, mock_filter, est_simple_enrollment_view) -> None:
+def test_get_credential_for_certificate(mock_get, mock_filter) -> None:
     """Test the get_credential_for_certificate method."""
     cert_mock = MagicMock()
     cert_mock.fingerprint.return_value = b'sample_fingerprint'
@@ -60,7 +61,7 @@ def test_get_credential_for_certificate(mock_get, mock_filter, est_simple_enroll
     mock_issued_credential.device = MagicMock()
     mock_get.return_value = mock_issued_credential
 
-    issued_credential = est_simple_enrollment_view.get_credential_for_certificate(cert_mock)
+    issued_credential = IssuedCredentialModel.get_credential_for_certificate(cert_mock)
     assert issued_credential == mock_issued_credential
     assert issued_credential.device == mock_issued_credential.device
     assert issued_credential.credential == mock_credential
@@ -111,71 +112,15 @@ def test_issue_credential_invalid_template(mock_issue, est_simple_enrollment_vie
             'invalid_template', MagicMock(), est_simple_enrollment_view.requested_domain, MagicMock()
         )
 
-# TLS client certificate verification tests
-
-def test_tls_client_cert_verification(est_simple_enrollment_view) -> None:
-    """Tests the TLS client certificate verification with the direct Issuing CA in the Truststore."""
-    certs, _keys = CertificateGenerator.create_test_pki(1)
-    truststore = get_mock_truststore([certs[0]])
-    assert est_simple_enrollment_view._verify_idevid_against_truststore(certs[1], [], truststore) # noqa: SLF001
-
-def test_tls_client_cert_verification_self_signed(est_simple_enrollment_view) -> None:
-    """Tests the TLS client certificate verification with a self-signed client certificate."""
-    certs, _keys = CertificateGenerator.create_test_pki(0)
-    truststore = get_mock_truststore([certs[0]])
-    assert est_simple_enrollment_view._verify_idevid_against_truststore(certs[0], [], truststore) # noqa: SLF001
-
-def test_tls_client_cert_verification_not_in_truststore(est_simple_enrollment_view) -> None:
-    """Tests the TLS client certificate verification fails if no matching certificate in the Truststore."""
-    different_pki_root, _key = CertificateGenerator.create_root_ca('Different Root CA')
-    truststore = get_mock_truststore([different_pki_root])
-    certs, _keys = CertificateGenerator.create_test_pki(1)
-    assert not est_simple_enrollment_view._verify_idevid_against_truststore(certs[1], [], truststore) # noqa: SLF001
-
 def test_tls_client_cert_verification_no_cert(est_simple_enrollment_view) -> None:
     """Tests the TLS client certificate verification if no valid PEM is passed."""
     est_simple_enrollment_view.request.META = {
         'SSL_CLIENT_CERT': '41foobar',
     }
     with pytest.raises(ClientCertificateAuthenticationError):
-        est_simple_enrollment_view.authenticate_idevid(
+        IDevIDAuthenticator.authenticate_idevid(
             est_simple_enrollment_view.request, est_simple_enrollment_view.requested_domain
         )
-
-@pytest.mark.parametrize('client_includes_root_ca', [True, False])
-def test_tls_client_cert_verification_chain(est_simple_enrollment_view, client_includes_root_ca: bool) -> None:  # noqa: FBT001
-    """Tests the TLS client certificate verification with an intermediate CA."""
-    # TODO(Air): This test will only work once the cryptography 44 hack is removed.
-    return
-    certs, _keys = CertificateGenerator.create_test_pki(2)
-    truststore = get_mock_truststore([certs[0]])
-    intermediates = [certs[1], certs[0]] if client_includes_root_ca else [certs[1]]
-    assert est_simple_enrollment_view._verify_idevid_against_truststore(certs[2], intermediates, truststore) # noqa: SLF001
-
-def test_tls_client_cert_chain_too_long(est_simple_enrollment_view) -> None:
-    """Tests the TLS client certificate verification with a too long chain."""
-    certs, _keys = CertificateGenerator.create_test_pki(4)
-    truststore = get_mock_truststore([certs[0]])
-    intermediates = [certs[3], certs[2], certs[1]]
-    assert not est_simple_enrollment_view._verify_idevid_against_truststore(certs[4], intermediates, truststore) # noqa: SLF001
-
-def test_tls_client_cert_idevid_in_truststore(est_simple_enrollment_view) -> None:
-    """Tests that verification works with just the IDevID directly in the Truststore (not the CA)."""
-    certs, _keys = CertificateGenerator.create_test_pki(1)
-    truststore = get_mock_truststore([certs[1]])
-    assert est_simple_enrollment_view._verify_idevid_against_truststore(certs[1], [], truststore) # noqa: SLF001
-
-def test_tls_client_cert_idevid_expired() -> None:
-    """Tests that verification fails if the IDevID certificate is expired."""
-
-
-def test_tls_client_cert_ca_expired() -> None:
-    """Tests that verification fails if the CA certificate is expired."""
-
-
-def test_tls_client_cert_attributes() -> None:
-    """Tests that verification fails if the client cert has no subject serial number."""
-
 
 def test_tls_client_cert_domain_credential_enrollment() -> None:
     """Tests that an issued credential can be enrolled via EST simpleenroll using an IDevID."""
