+++ trustpoint/templates/devices/devices_table.html
@@ -23,7 +23,7 @@
             {% endif %}
             <th>
                 <a href="?sort={% if current_sort == 'created_at' %}-{% endif %}created_at">
-                    {% trans 'Created' %}
+                    {% trans 'Created at' %}
                 </a>
             </th>
             <th>
